Add win16/win32 STRICT syntax to all sdk projects. Some corrections to sdk projects.
[git-p4: depot-paths = "//depot/openwatcom/": change = 21821]

+++ bld/dig/c/win/dipld.c
@@ -97,7 +97,7 @@ dip_status DIPSysLoad( char *path, dip_client_routines *cli,
     dll = LoadModule( newpath, &parm_block );
     DIPLastHandle = dll;
     SetErrorMode( prev );
-    if( dll < 32 ) {
+    if( (UINT)dll < 32 ) {
         return( DS_ERR|DS_FOPEN_FAILED );
     }
     *sys_hdl = (unsigned long)transfer_block.unload;
